+++ src/assets/scripts/main.js
@@ -3,6 +3,7 @@
 // require('./module/rAF');
 var fastclick = require('fastclick');
 // var _ = require('./helpers/helpers');
+var Layzr = require('layzr.js');
 
 var _init = false
 
@@ -24,6 +25,10 @@ var setTheme = function() {
   }
 }
 
+var bindLazyLoad = function(option) {
+  var layzr = new Layzr(option);
+}
+
 var app = {
 
   init: function() {
@@ -31,6 +36,11 @@ var app = {
     _init = true
 
     setTheme()
+    bindLazyLoad({
+      callback: function() {
+        this.parentElement.classList.add('loaded');
+      }
+    })
 
     // Attach fastclick
     fastclick.attach(document.body);