+++ CHANGELOG.md
@@ -15,3 +15,14 @@
 - Fixed: Google Analytics jekyll production ENV 
 - Added: rtl CSS helper class
 - Fixed: missing drafts folder 
+
+# v1.5.0
+
+- Added: jekyll figure tag plugin
+- Added: jekyll aparat video tag plugin
+- Added: lazy load to image and iframe
+- Added: social sharing card image mata tag
+- Updated: some css typography styles
+- Update: post summary style and markup
+- Fixed: list item margin bottom style in article
+- Fixed: youtube Tag plugin markup (responsive)